requirements.txt
fastapi
uvicorn
requests
pillow
opencv-python
numpy
faiss-cpu
transformers
torch
ultralytics